feat: remove pinia dependency and add full bundle
