+++ packages/@n8n/chat/src/composables/index.ts
@@ -1,2 +1,3 @@
+export * from './useChat';
 export * from './useI18n';
 export * from './useOptions';
