+++ packages/@n8n/chat/src/composables/useChat.ts
@@ -1,0 +1,7 @@
+import { inject } from 'vue';
+import { ChatSymbol } from '@/constants';
+import type { Chat } from '@/types';
+
+export function useChat() {
+	return inject(ChatSymbol) as Chat;
+}
