+++ packages/@n8n/chat/src/plugins/index.ts
@@ -1,0 +1,1 @@
+export * from './chat';
