+++ packages/@n8n/chat/src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './chat';
 export * from './messages';
 export * from './options';
 export * from './webhook';
