+++ packages/@n8n/chat/vite.config.ts
@@ -6,6 +6,8 @@ import vue from '@vitejs/plugin-vue';
 import icons from 'unplugin-icons/vite';
 import dts from 'vite-plugin-dts';
 
+const includeVue = process.env.INCLUDE_VUE === 'true';
+
 // https://vitejs.dev/config/
 export default defineConfig({
 	plugins: [
@@ -21,15 +23,16 @@ export default defineConfig({
 		},
 	},
 	build: {
+		emptyOutDir: !includeVue,
 		lib: {
 			entry: resolve(__dirname, 'src', 'index.ts'),
 			name: 'N8nChat',
-			fileName: (format) => `chat.${format}.js`,
+			fileName: (format) => (includeVue ? `chat.bundle.${format}.js` : `chat.${format}.js`),
 		},
 		rollupOptions: {
 			// make sure to externalize deps that shouldn't be bundled
 			// into your library
-			external: ['vue'],
+			external: includeVue ? [] : ['vue'],
 			output: {
 				exports: 'named',
 				// Provide global variables to use in the UMD build
